fix chaos connect script
Previous the created ip route was not deleted, since the 'command ip' check was also not tested in the pod more on the machine where the experiment was running. Furthermore it was problematic if a leader change happened since we only tried to reset for the current leader for the partitions. We no try to remove the ip route on all nodes if ip is installed