fix: resolve files correctly in jar
We need to use here a spring helper class, in order to resolve the files inside a
jar. In unit tests it simply works with  but at the end in the jar
not anymore, because the jar has to be handled like a ZIP.
See https://stackoverflow.com/questions/51972708/listfiles-method-in-java-doesnt-work-in-jar
If we directly address the file and resolve it is not an issue, but for listing files it is.