feat: new subcmd deploy process models
Add new sub command to deploy process models, either specified by the user
via processModelPath or if not given it default process models BPMN and DMN
are be deployed.

The user is able to specify the count of versions which should be deployed for each.
This allows to experiment with deployment distribution.