refactor: iterate over steady-state cmd
Rename command to instance-creation, because this it is what it does or should do.
Verifying whehter instance creation works on a specific partition for a specfic model.

Remove deployment from command, and the processDefinitionKey usage.
It is unlikely that we use this in our experiments it is easier to use the named
process id and version (if not set we use the latest version).