fix: handle worker deployment in self-managed env
In self-managed env the zeebe gateway service is called differently, based on the helm release name.
The release name is in our setups normally also the namespace name, which we use here.

If the worker is already deployed we update the deployment, instead of failing.

Added missing tests for it.