fix: only delete pvc
Reclaim policy of pvc is "Delete", which means it will be automatically deleted when PV is deleted and it is no longer attached to a pod.
When we delete both PVC and PV together, the disks are not getting deleted. We don't know yet why this happens. But we verified that when we delete PVC, the PV is also getting deleted.